use less janky gamepad cursor positioning

+++ include/mouse.h
@@ -138,7 +138,7 @@ void mouse_get_delta(int *dx = NULL, int *dy = NULL, int *dz = NULL);
 void mouse_force_pos(int x, int y);
 void mouse_update_pos(float x, float y, float dx, float dy);
 // update mouse with position which is already scaled for max_w/max_h
-void mouse_update_pos_scaled(int x, int y, int dx, int dy);
+void mouse_update_pos_scaled(int x, int y, float dx, float dy);
 
 #endif
